Ship max Kimi reasoning by default
