+++ trigger/agent-long.ts
@@ -117,7 +117,10 @@ import {
   createChatLogger,
   type ChatLogger,
 } from "@/lib/api/chat-logger";
-import { evaluateKimiReasoningExperiment } from "@/lib/experiments/kimi-reasoning";
+import {
+  KIMI_MAX_REASONING_EFFORT,
+  shouldUseMaxKimiReasoning,
+} from "@/lib/ai/kimi-reasoning";
 import {
   LEGACY_AGENT_API_ENDPOINT,
   type AgentApiEndpoint,
@@ -2522,15 +2525,12 @@ export const agentLongTask = task({
             const streamStartTime = taskStartTime;
             const configuredModelId =
               trackedProvider.languageModel(selectedModel).modelId;
-            const kimiReasoningExperiment =
-              await evaluateKimiReasoningExperiment({
-                posthog,
-                userId,
-                subscription,
-                mode,
-                selectedModel,
-                configuredModelId,
-              });
+            const useMaxKimiReasoning = shouldUseMaxKimiReasoning({
+              subscription,
+              mode,
+              selectedModel,
+              configuredModelId,
+            });
             const budgetMonitor = effectiveBudgetSnapshot
               ? new BudgetMonitor(
                   effectiveBudgetSnapshot,
@@ -2758,7 +2758,6 @@ export const agentLongTask = task({
                   analyticsRequestContext,
                   usage: usageCostRecord,
                   responseModel: state.responseModel,
-                  kimiReasoningExperiment,
                   ...(usageSettlementState && {
                     usageSettlement: {
                       id: usageTracker.usageSettlementId,
@@ -2933,12 +2932,12 @@ export const agentLongTask = task({
                 }
               },
               settleUsageAfterStep,
-              ...(kimiReasoningExperiment && {
+              ...(useMaxKimiReasoning && {
                 providerReasoningOverride: {
                   modelName: selectedModel,
                   reasoning: {
                     enabled: true,
-                    effort: kimiReasoningExperiment.reasoningEffort,
+                    effort: KIMI_MAX_REASONING_EFFORT,
                   },
                 },
               }),
@@ -3309,7 +3308,6 @@ export const agentLongTask = task({
                                       state.budgetAbortDetails,
                                     agentPermissionMode,
                                     isAutoContinue: !!isAutoContinue,
-                                    kimiReasoningExperiment,
                                     stepLimitTelemetry:
                                       buildAgentStepLimitTelemetry({
                                         configuredMaxSteps:
@@ -3472,7 +3470,6 @@ export const agentLongTask = task({
                         budgetAbortDetails: state.budgetAbortDetails,
                         agentPermissionMode,
                         isAutoContinue: !!isAutoContinue,
-                        kimiReasoningExperiment,
                         stepLimitTelemetry: buildAgentStepLimitTelemetry({
                           configuredMaxSteps: state.configuredMaxSteps,
                           stepCount: state.agentStepCount,
